pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "groq>=1.1.2",
    "joblib>=1.5.3",
    "numpy>=2.4.4",
    "openai>=2.31.0",
    "pandas>=3.0.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.24",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
    "xgboost>=3.2.0",
]
